fix: only check for `self.ws` while holding the race lock
This was a really hard bug to track down...

What ended up happening was that the WebSocket would be reconnecting at the exact moment the heartbeater was trying to send a heartbeat.
While the gateway was designed to handle this, an oversight was made that checked if `self.ws` was None while not holding the lock required to rely on `self.ws`.

To help resolving future issues, an extra `run_bee_gees()` method is added that immediately closes the gateway if the heartbeater raises an error.