fix: Restore previous wait code
The new code utilised asyncio Events. This broke waits. 

For one, `asyncio.wait` does not support `asyncio.Event`, instead you should use `asyncio.Event.wait()`
Secondly, `client.wait_for` is supposed to return the data it is waiting for, even after fixing the first issue, it would simply return a boolean, which is not what we want 

If necessary, someone is welcome to submit a PR that uses Events, but i dont believe there is a benefit in converting