feat: add `unavailable` bool to `GuildUnavailable`
Sometimes, an "Unavailable Guild" object returned from the API can also return an `unavailable` boolean denoting whether it's accessible or not. For clarification, this was internally explained that this is a value given for events such as guild termination via. TnS and can represent whether it's actually unavailable by means of the server being deleted, or if the `GUILD_CREATE` event was unable to return data at the time of its dispatch for whatever API issues may be present.