feat💥: convert `cover_image` in `Application` to `Asset`. (#475)
The `cover_image` is currently handled as a string of its image hash. Since this attribute in the Discord Developers documentation shares the [same CDN endpoint as the `icon` attribute,](https://i.imgur.com/G0fXSfq.png) I've made a quick edit to its typing and altered the `_process_dict` classmethod to reflect it.